// src/types.ts
export interface SiteConfig {
  author: string;
  date: {
    locale: string | string[] | undefined;
    options: Intl.DateTimeFormatOptions;
  };
  description: string;
  lang: string;
  ogLocale: string;
  title: string;
  subTitle: string;
  url: string;
}

export interface PaginationLink {
  srLabel?: string;
  text?: string;
  url: string;
}

export interface SiteMeta {
  articleDate?: string | undefined;
  description?: string;
  ogImage?: string | undefined;
  title: string;
}
export interface Author {
  name: string;
  photo: string;
  type: string;
  url: string;
}

export interface Content {
  "content-type": string;
  html: string;
  text: string;
  value: string;
}

export interface Rels {
  canonical: string;
}

export interface Summary {
  "content-type": string;
  value: string;
}

export type AdmonitionType =
  | "seedling"
  | "blossoming"
  | "branching"
  | "flowering"
  | "fruit"
  | "published"
  | "tip"
  | "note"
  | "important"
  | "caution"
  | "warning";